# docs/swagger.yaml
definitions:
...
  model.ResetEmailRequest:
    properties:
      email:
        maxLength: 255
        type: string
      tokenCaptcha:
        minLength: 100
        type: string
    required:
    - email
    - tokenCaptcha
    type: object
  model.ResetRequest:
    properties:
      code:
        maxLength: 255
        minLength: 36
        type: string
      confirmPassword:
        type: string
      password:
        maxLength: 255
        minLength: 8
        type: string
    required:
    - code
    - confirmPassword
    - password
    type: object
...
  utility.ResponseError:
    properties:
      error:
        type: string
    type: object
  utility.ResponseSuccess:
    properties:
      data: {}
    type: object
info:
  contact: {}
paths:
...
  /api/reset/email:
    post:
      consumes:
      - application/json
      description: Sends an email containing a password reset link
      parameters:
      - description: User email for password reset
        in: body
        name: resetEmail
        required: true
        schema:
          $ref: '#/definitions/model.ResetEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utility.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utility.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utility.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utility.ResponseError'
      summary: Request password reset
      tags:
      - Reset
  /api/reset/password:
    post:
      consumes:
      - application/json
      description: Resets the user's password using the provided reset code and new
        password
      parameters:
      - description: Request containing reset code and new password
        in: body
        name: resetRequest
        required: true
        schema:
          $ref: '#/definitions/model.ResetRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utility.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utility.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utility.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utility.ResponseError'
      summary: Reset password
      tags:
      - Reset
...
swagger: "2.0"